agcms-gateway/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.4
httpx==0.27.0
redis==5.0.3
pyyaml==6.0.1
asyncpg==0.29.0
python-jose[cryptography]==3.3.0
